+++ src/main/java/com/zhao/mawen/model/Comment.java
@@ -3,17 +3,16 @@ package com.zhao.mawen.model;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.util.Arrays;
 
 /**
  * comment
  * @author 
  */
 @Data
-public class Comment{
+public class Comment {
     private Long id;
 
-    private Integer parentId;
+    private Long parentId;
 
     private Integer type;
